Refactor PasswordResetActivity & VerificationActivity - Cleanup strings - Edit Styles & Colors - Update Manifest info
