+++ app/src/main/java/org/health17/iwatch/PasswordResetActivity.kt
@@ -1,10 +1,7 @@
 package org.health17.iwatch
 
-import android.content.Context
-import android.content.Intent
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.widget.ImageView
+import androidx.appcompat.app.AppCompatActivity
 import kotlinx.android.synthetic.main.activity_password_reset.*
 
 class PasswordResetActivity : AppCompatActivity() {
@@ -12,13 +9,12 @@ class PasswordResetActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_password_reset)
 
-        view_pager.adapter = TabAdapter(this, supportFragmentManager)
-        tab_layout.setupWithViewPager(view_pager)
+        setSupportActionBar(password_reset_toolbar)
+        supportActionBar?.setDisplayShowTitleEnabled(false)
+        password_reset_toolbar.setNavigationOnClickListener { finish() }
 
-        val back = findViewById<ImageView>(R.id.back)
-        back.setOnClickListener{
-            val intent = Intent(this, VerificationActivity::class.java)
-            startActivity(intent)
-        }
+        password_reset_viewpager.adapter = TabAdapter(this, supportFragmentManager)
+        password_reset_tabLayout.setupWithViewPager(password_reset_viewpager)
+
     }
 }
