+++ app/src/main/java/org/health17/iwatch/TabAdapter.kt
@@ -4,10 +4,10 @@ import android.content.Context
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.FragmentPagerAdapter
-import org.health17.iwatch.ui.main.email_verification.EmailVerificationFragment
-import org.health17.iwatch.ui.main.mobile_verification.MobileVerificationFragment
+import org.health17.iwatch.fragments.email_verification.EmailVerificationFragment
+import org.health17.iwatch.fragments.mobile_verification.MobileVerificationFragment
 
-class TabAdapter (private val mContext: Context, fragmentManager: FragmentManager) :
+class TabAdapter(private val mContext: Context, fragmentManager: FragmentManager) :
     FragmentPagerAdapter(fragmentManager) {
     override fun getItem(position: Int): Fragment {
         return when (position) {
